Camera Scrolling & Map tile & Bug fix  On branch gameStudy  Your branch is up to date with 'origin/gameStudy'.  Changes to be committed: 	modified:   TCPfighter/TCPfighter/Animation.cpp 	modified:   TCPfighter/TCPfighter/CGameBase.h 		-> Add ScreenPos
	new file:   TCPfighter/TCPfighter/Camera.cpp
	new file:   TCPfighter/TCPfighter/Camera.h
		-> Add Camera : FreeCam/TraceCam mode support

	modified:   TCPfighter/TCPfighter/Constants.cpp
	modified:   TCPfighter/TCPfighter/Constants.h
	modified:   TCPfighter/TCPfighter/DataStructures.h
	modified:   TCPfighter/TCPfighter/GameManager.cpp
	modified:   TCPfighter/TCPfighter/GameManager.h
	modified:   TCPfighter/TCPfighter/GameNetReactor.cpp
	modified:   TCPfighter/TCPfighter/GameSystemInfo.cpp
	modified:   TCPfighter/TCPfighter/HpGuage.cpp

	modified:   TCPfighter/TCPfighter/ImageProcessor.cpp
	modified:   TCPfighter/TCPfighter/ImageProcessor.h
		-> Clipping & AlphaBlending bug fix

	new file:   TCPfighter/TCPfighter/Map.cpp
	new file:   TCPfighter/TCPfighter/Map.h
		-> Add Map : Tile 100 x 100
		<Need Performance Improvement>

	modified:   TCPfighter/TCPfighter/Player.cpp
	modified:   TCPfighter/TCPfighter/Shadow.cpp
	modified:   TCPfighter/TCPfighter/TCPfighter.cpp
	modified:   TCPfighter/TCPfighter/TCPfighter.vcxproj
	modified:   TCPfighter/TCPfighter/TCPfighter.vcxproj.filters
	modified:   TCPfighter/TCPfighter/stdafx.h

+++ TCPfighter/TCPfighter/Animation.cpp
@@ -35,7 +35,7 @@ AnimStruct* Animation::MakeAnimationStruct(KsDIB* image, INT frameDelay, Positio
 
 bool Animation::AddSprite(AnimStruct* sprite, bool bOverride)
 {
-	return AddSprite(sprite, m_anim.size(), bOverride);
+	return AddSprite(sprite, (int)m_anim.size(), bOverride);
 }
 
 bool Animation::AddSprite(AnimStruct* sprite, INT idx, bool bOverride)
